Evaluate command as enter when text equals delimiter
A document command represents a text modification as a document replace command. The length of a document command corresponds to the length of the selection region when pasting into the editor. The document command text corresponds to the clipboard after copying. This method incorrectly uses these values to evaluate if the document command was a user pressing the enter key. The enter key check evaluates true under the following conditions: length is 0, text is not null, and text ends with a delimiter. The bug is when someone copies text like “hello\n” and pastes it. This will be considered as the enter key being pressed, because the text ends with a delimiter. However, we know the document command was not someone pressing the enter key, as it was text that happened to end with a delimiter.