fix: return correct value of "dark" property for default Eclipse theme
On a new workspace, prefs.get(PreferenceConstants.E4_THEME_ID, null)
will return null, because no theme changes have yet taken place.

In this case (i.e., when the default theme is active), it is sensible to
default to the system's dark mode setting, instead of defaulting to the
light mode.